# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    - sku
    - stock
    type: object
...
  github_com_NR3101_go-ecom-project_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse'
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    type: object
...
  github_com_NR3101_go-ecom-project_internal_dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    - stock
    type: object
...
  github_com_NR3101_go-ecom-project_internal_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.PaginationMeta'
      success:
        type: boolean
    type: object
  github_com_NR3101_go-ecom-project_internal_utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_NR3101_go-ecom-project_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: no-email@enail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for an e-commerce application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swaager.io/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve all active product categories
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category (Admin only)
      parameters:
      - description: Category creation data
        in: body
        name: createCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing product category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update data
        in: body
        name: updateCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Invalid request body or category ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve all products with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (Admin only)
      parameters:
      - description: Product creation data
        in: body
        name: createProductRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update data
        in: body
        name: updateProductRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid request body or product ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /products/{id}/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Product image uploaded successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid product ID or image file required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Upload a product image
      tags:
      - Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate requests
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"